Add in old toggle switches
Lets face it, the old toggle switches had a nice feel to them. I have
reverted the checkboxes, however rewrote the way their backend is
structured so it is a bit more logical. Instead of being represented by
a counter, it is being represented by a boolean, which is simply
inverted whenever the switch is clicked.